Lesson6_HomeWork_Stage1
The animation of the avatar was added.
