+++ SwiftUIVK/SwiftUIVK/Views/FriendsView.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import SystemConfiguration
 
 struct FriendsView: View {
     
@@ -21,6 +22,7 @@ struct FriendsView: View {
             List(session.friends) { friend in
                 FriendCell(friend: friend, isPhotoButtonSelected: $shouldShowPhotoCollection)
             }.navigationBarTitle(Text("Friends"))
+            
             
             NavigationLink(destination: PhotoCollectionView(), isActive: $shouldShowPhotoCollection) {
                 EmptyView()
@@ -44,37 +46,4 @@ struct FriendsView_Previews: PreviewProvider {
 }
 
 
-struct FriendCell: View {
-    
-    let friend: Friend
-    
-    @Binding var isPhotoButtonSelected: Bool
-    
-    var body: some View {
-        
-        HStack {
-            ImageBuilder {
-                Image(uiImage: ImageLoader().getImage(friend.photo50))
-            }
-            
-            VStack(alignment: .leading) {
-                Text("\(friend.firstName) \(friend.lastName)")
-                    .font(.subheadline)
-                    .foregroundColor(.black)
-                Text("id: \(String(friend.id))")
-                    .font(.subheadline)
-                    .foregroundColor(.gray)
-            }.padding(.leading, 30)
-            
-            
-            Button {
-                isPhotoButtonSelected = true
-            } label: {
-                Text("")
-                    .font(.subheadline)
-                    .foregroundColor(.blue)
-                    .padding(.leading, 30)
-            }
-        }
-    }
-}
+
